+++ src/dialogs/clipboardhistorydialog.cpp
@@ -360,7 +360,7 @@ void ClipboardHistoryDialog::updateViewGeometry()
     const QSize centralSize = m_mainWindow->size();
 
     // width: 2.4 of editor, height: 1/2 of editor
-    const QSize viewMaxSize(centralSize.width() / 2.4, centralSize.height() / 2);
+    const QSize viewMaxSize(static_cast<int>(centralSize.width() / 2.4), centralSize.height() / 2);
 
     // Position should be central over window
     const int xPos = std::max(0, (centralSize.width() - viewMaxSize.width()) / 2);
